perf: f16 KV borrow attention + stream-restart prefetch -> 0.95x of ollama
Three decode fixes on top of the fused-region work:

- f16 KV cache borrow path: new KvElem trait makes the online-softmax
  decode kernel generic over the KV element; u16 rows convert in-kernel
  via F16C (AVX2), f32 passes through bit-identically. The KV cache gains
  f16_layer_{key,value}_prefix borrows, and decode attention prefers them
  before the f32 borrow / dequant-copy paths. `run`/`serve` KV default
  f32 -> f16: zero-copy like f32 but half the attention DRAM reads as the
  context grows, and half the memory.
- Next-quad prefetch sweep in the Q4_K/Q6_K x4 row kernels for short rows
  (blocks_per_row <= 16): 10-block rows restart the hardware prefetcher
  every 22 cache lines, which held every 2560-column matrix (gate/up
  projections, q/k/v, lm_head) ~10-20% under the DRAM roof. The sweep
  walks the next quad's row one quad-time ahead: gate/up 37 -> 45.6 GB/s,
  qkv 34 -> 41.9, lm_head 38 -> 44.3, decode 13.3 -> 15.1 tok/s
  (same-conditions A/B). Long rows get a deeper in-row T1 sweep instead
  (down-proj relative deficit closed in the shape microbench).
- gemm_quantized_f32 now records into the OXIDIZE_DECODE_PROFILE summary
  (prefill attribution; batch-43 GEMM measured ~46% of FMA peak).

Benchmarks (Ryzen 6850H, Qwen3-4B Q4_K_M, cool machine, same-run pairs):
  decode-only token_forward: 70.1 -> 62.7 ms/token (15.95 tok/s)
  oxidize self-reported 512 tok: 13.3 -> 15.1-15.2 tok/s
  ollama-performance-benchmark (768 tok, load included, ollama runs
  first on the cooler machine): oxidize 14.91 vs ollama 15.72 = 0.95x
  (was 0.61x at the start of this effort)

f16 attention matches f32 within half-precision rounding (test included).

+++ oxidize-cli/src/main.rs
@@ -698,13 +698,13 @@ where
         }
     }
     if !has_flag(&rewritten, "--kv-cache-dtype") {
-        // f32 is the only KV dtype the decode attention path can borrow
-        // zero-copy; q8/f16 dequantize the WHOLE K/V prefix into workspace
-        // buffers every layer, every token. cpu-optimized clamps the context
-        // to 2048, bounding the f32 cache (~600 MB for a 4B model). Pass
+        // f16/f32 are the KV dtypes decode attention can borrow zero-copy
+        // (f16 converts in-kernel via F16C); q8 dequantizes the WHOLE K/V
+        // prefix into workspace buffers every layer, every token. f16 also
+        // halves attention DRAM reads vs f32 as the context grows. Pass
         // --kv-cache-dtype q8 to trade decode speed for memory.
         rewritten.push("--kv-cache-dtype".into());
-        rewritten.push("f32".into());
+        rewritten.push("f16".into());
     }
     // One-shot prompt runs exit right after generation, so a background API
     // server would just load the model a second time (concurrently, stealing
@@ -808,11 +808,10 @@ fn rewrite_serve_args(raw: Vec<OsString>) -> io::Result<Vec<OsString>> {
         rewritten.push(model_path.into_os_string());
     }
     if !has_flag(&rewritten, "--kv-cache-dtype") {
-        // Match the `run` rewrite: f32 KV is the zero-copy decode path (see
-        // the comment there); the server's ctx auto-cap accounts for the
-        // larger per-token KV footprint.
+        // Match the `run` rewrite: f16 KV is the zero-copy decode path with
+        // half the attention reads of f32 (see the comment there).
         rewritten.push("--kv-cache-dtype".into());
-        rewritten.push("f32".into());
+        rewritten.push("f16".into());
     }
     if !has_flag(&rewritten, "--cpu-optimized") {
         rewritten.push("--cpu-optimized".into());
@@ -2879,7 +2878,7 @@ mod tests {
         assert!(args.contains(&OsString::from("--mmap-prefetch")));
         assert!(args.contains(&OsString::from("--mmap-hugepages")));
         assert!(args.contains(&OsString::from("--kv-cache-dtype")));
-        assert!(args.contains(&OsString::from("f32")));
+        assert!(args.contains(&OsString::from("f16")));
     }
 
     #[test]
